UCAAS-1486: move TypeScript gluecode tests to ts-gen/tests
Keep gluecode limited to compiler-emitted runtime files; run scripts and build docs point at the dedicated test folder.
