+++ compiler/back-ends/ts-gen/tests/TSASN1Base.remoteCapability.test.ts
@@ -1,19 +1,19 @@
-// Run: npx tsx compiler/back-ends/ts-gen/gluecode/TSASN1Base.remoteCapability.test.ts
+// Run: npx tsx compiler/back-ends/ts-gen/tests/TSASN1Base.remoteCapability.test.ts
 import assert from "node:assert/strict";
 import test from "node:test";
 import {
 	ASN1ClassInstanceType,
 	TSASN1Base,
-} from "./TSASN1Base.js";
-import { EASN1TransportEncoding } from "./TSInvokeContext.js";
+} from "../gluecode/TSASN1Base.js";
+import { EASN1TransportEncoding } from "../gluecode/TSInvokeContext.js";
 import {
 	CustomInvokeProblemEnum,
 	RemoteCapabilityMode,
 	ROSE_REJECT_REMOTENOTCAPABLE,
-} from "./TSROSEBase.js";
-import { buildRemoteModuleCapabilities } from "./TSModuleCapabilities.js";
-import type { IASN1InvokeData } from "./TSROSEBase.js";
-import type { ROSEError, ROSEInvoke, ROSEReject, ROSEResult } from "./SNACCROSE.js";
+} from "../gluecode/TSROSEBase.js";
+import { buildRemoteModuleCapabilities } from "../gluecode/TSModuleCapabilities.js";
+import type { IASN1InvokeData } from "../gluecode/TSROSEBase.js";
+import type { ROSEError, ROSEInvoke, ROSEReject, ROSEResult } from "../gluecode/SNACCROSE.js";
 
 class TestTransport extends TSASN1Base {
 	public sendInvokeCount = 0;
