+++ scripts/run_gluecode_tests.bat
@@ -4,6 +4,7 @@ setlocal EnableExtensions
 set "SCRIPT_DIR=%~dp0"
 set "REPO_ROOT=%SCRIPT_DIR%.."
 set "GLUE_DIR=%REPO_ROOT%\compiler\back-ends\ts-gen\gluecode"
+set "TEST_DIR=%REPO_ROOT%\compiler\back-ends\ts-gen\tests"
 set "STUB_DIR=%REPO_ROOT%\samples\ts-microservice\node-client\src\stub"
 set "NODE_MODULES=%REPO_ROOT%\samples\ts-microservice\node-client\node_modules"
 
@@ -28,7 +29,7 @@ for %%T in (
 	TSModuleCapabilities.test.ts
 ) do (
 	echo Running %%T ...
-	npx --yes tsx "%GLUE_DIR%\%%T"
+	npx --yes tsx "%TEST_DIR%\%%T"
 	if errorlevel 1 set "EXIT_CODE=1"
 )
